+++ src/main/java/com/xw/framework/domain/IErrorCode.java
@@ -8,7 +8,7 @@ public interface IErrorCode {
     /**
      * 返回码
      */
-    long getCode();
+    int getCode();
 
     /**
      * 返回信息
